:pencil: Add jsDoc to inspecteur-queries
