+++ server/src/models/place/place.queries.js
@@ -5,7 +5,6 @@
 import mongoose from 'mongoose'
 
 import Place from './place.model'
-import '../inspecteur/inspecteur.model'
 import { appLogger, techLogger } from '../../util'
 import { createArchivedPlaceFromPlace } from '../archived-place/archived-place-queries'
